Jump flappy bird + gravity + basic collision (up and down)

+++ 2D_Game/2D_Game/Display.cpp
@@ -25,7 +25,7 @@ void Display::loadUnit(float x, float y, AUnit* unit)
   sf::Sprite	Sprite;
 
   Sprite.setTexture(unit->texture);
-  Sprite.setPosition(unit->x, unit->y);
+  Sprite.setPosition((float)unit->x, (float)unit->y);
   Sprite.setTextureRect(sf::IntRect(unit->width * unit->animFrame + (3 * unit->width * unit->act), unit->height * unit->dir, unit->width, unit->height));
   win.draw(Sprite);
 }
